need to coerce mbscValid to signed type
git-svn-id: https://svn.r-project.org/R/trunk@32954 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/util.c
@@ -1167,7 +1167,7 @@ size_t Mbrtowc(wchar_t *wc, const char *s, size_t n, mbstate_t *ps)
 
 Rboolean mbcsValid(char *str)
 {
-    return  (mbstowcs(NULL, str, 0) >= 0);
+    return  ((int)mbstowcs(NULL, str, 0) >= 0);
 }
 
 /* We do this conversion ourselves to do our own error recovery */
